[firebase_ml_vision] v2 embedding API (#275)
