+++ packages/firebase_ml_vision/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.9.3
+
+* Support v2 embedding. This plugin will remain compatible with the original embedding and won't
+require app migration.
+
 ## 0.9.2+3
 
 * Use `BoxDecoration` `const` constructor in example app.
