+++ packages/firebase_ml_vision/android/gradle.properties
@@ -1,3 +1,1 @@
 org.gradle.jvmargs=-Xmx1536M
-android.enableJetifier=true
-android.useAndroidX=true
