+++ packages/firebase_ml_vision/example/android/app/build.gradle
@@ -49,6 +49,9 @@ android {
     dependencies {
         api 'com.google.firebase:firebase-ml-vision-image-label-model:17.0.2'
         api 'com.google.firebase:firebase-ml-vision-face-model:17.0.2'
+        androidTestImplementation 'androidx.test:runner:1.2.0'
+        androidTestImplementation 'androidx.test:rules:1.2.0'
+        androidTestImplementation 'androidx.test.espresso:espresso-core:3.2.0'
     }
 }
