+++ packages/firebase_ml_vision/example/android/app/src/androidTest/java/io/flutter/plugins/firebasemlvision/EmbeddingV1ActivityTest.java
@@ -1,0 +1,14 @@
+package io.flutter.plugins.firebasemlvision;
+
+import androidx.test.rule.ActivityTestRule;
+import dev.flutter.plugins.e2e.FlutterRunner;
+import io.flutter.plugins.firebasemlvisionexample.EmbeddingV1Activity;
+import org.junit.Rule;
+import org.junit.runner.RunWith;
+
+@RunWith(FlutterRunner.class)
+public class EmbeddingV1ActivityTest {
+  @Rule
+  public ActivityTestRule<EmbeddingV1Activity> rule =
+      new ActivityTestRule<>(EmbeddingV1Activity.class);
+}
