+++ packages/firebase_ml_vision/example/android/app/src/androidTest/java/io/flutter/plugins/firebasemlvision/MainActivityTest.java
@@ -1,0 +1,12 @@
+package io.flutter.plugins.firebasemlvision;
+
+import androidx.test.rule.ActivityTestRule;
+import dev.flutter.plugins.e2e.FlutterRunner;
+import io.flutter.plugins.firebasemlvisionexample.MainActivity;
+import org.junit.Rule;
+import org.junit.runner.RunWith;
+
+@RunWith(FlutterRunner.class)
+public class MainActivityTest {
+  @Rule public ActivityTestRule<MainActivity> rule = new ActivityTestRule<>(MainActivity.class);
+}
