+++ packages/firebase_ml_vision/example/android/gradle.properties
@@ -1,3 +1,4 @@
 org.gradle.jvmargs=-Xmx1536M
 android.enableJetifier=true
 android.useAndroidX=true
+android.enableR8=true
