+++ packages/firebase_ml_vision/example/pubspec.yaml
@@ -11,9 +11,10 @@ dependencies:
   firebase_ml_vision:
     path: ../
   firebase_core: ^0.4.0
-  camera: ^0.5.2+1
+  camera: ^0.5.6+1
 
 dev_dependencies:
+  e2e: ^0.2.1
   flutter_test:
     sdk: flutter
   flutter_driver:
