+++ packages/firebase_ml_vision/pubspec.yaml
@@ -2,13 +2,14 @@ name: firebase_ml_vision
 description: Flutter plugin for Firebase machine learning vision services.
 author: Flutter Team <flutter-dev@googlegroups.com>
 homepage: https://github.com/FirebaseExtended/flutterfire/tree/master/packages/firebase_ml_vision
-version: 0.9.2+3
+version: 0.9.3
 
 dependencies:
   flutter:
     sdk: flutter
 
 dev_dependencies:
+  e2e: ^0.2.1
   image_picker: ^0.5.0
   flutter_test:
     sdk: flutter
@@ -27,4 +28,4 @@ flutter:
 
 environment:
   sdk: ">=2.0.0-dev.28.0 <3.0.0"
-  flutter: ">=1.5.0 <2.0.0"
+  flutter: ">=1.9.1+hotfix.5 <2.0.0"
